+++ CHANGELOG.md
@@ -122,6 +122,13 @@ Those changes in added, changed or breaking changes, should include usage exampl
         .await
         .unwrap();
   ```
+  
+* Added new method `set_ejector` in `avsregistry/writer` in [#330](https://github.com/Layr-Labs/eigensdk-rs/pull/330).
+  
+  ```rust
+    let new_ejector_address = address!("70997970C51812dc3A010C7d01b50e0d17dc79C8");
+    let tx_hash = avs_writer.set_ejector(new_ejector_address).await.unwrap();
+  ````
 
 * Added new method `eject_operator` in `avsregistry/writer` in [#328](https://github.com/Layr-Labs/eigensdk-rs/pull/328).